Rename some passes
