+++ src/opt/CMakeLists.txt
@@ -42,14 +42,14 @@ target_sources(${PROJECT_NAME}
         "ImproveMemoryOps.h"
         "InstructionSimplify.cpp"
         "InstructionSimplify.h"
-        "IntrinsicsOpt.cpp"
-        "IntrinsicsOpt.h"
+        "IntrinsicsOptPass.cpp"
+        "IntrinsicsOptPass.h"
         "MangleOpenCLBuiltins.cpp"
         "MangleOpenCLBuiltins.h"
         "MakeInternalFuncsStatic.cpp"
         "MakeInternalFuncsStatic.h"
-        "Peephole.cpp"
-        "Peephole.h"
+        "PeepholePass.cpp"
+        "PeepholePass.h"
         "ReplacePseudoMemoryOps.cpp"
         "ReplacePseudoMemoryOps.h"
         "ReplaceStdlibShiftPass.cpp"
