+++ src/opt/ISPCPasses.h
@@ -41,11 +41,11 @@
 #include "GatherCoalescePass.h"
 #include "ImproveMemoryOps.h"
 #include "InstructionSimplify.h"
-#include "IntrinsicsOpt.h"
+#include "IntrinsicsOptPass.h"
 #include "IsCompileTimeConstant.h"
 #include "MakeInternalFuncsStatic.h"
 #include "MangleOpenCLBuiltins.h"
-#include "Peephole.h"
+#include "PeepholePass.h"
 #include "ReplacePseudoMemoryOps.h"
 #include "ReplaceStdlibShiftPass.h"
 #include "XeGatherCoalescePass.h"
